Front_back_tumores/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pillow==10.3.0
numpy==1.26.4
joblib==1.3.2
scikit-learn==1.7.1
python-multipart==0.0.9
scikit-image==0.24.0
